Payment data type change and db lookup table

+++ Repository/DTO/ResponseDTO/Invoice/InvoiceItemView.cs
@@ -11,8 +11,8 @@ namespace Repository.DTO.ResponseDTO.Invoice
         public string Item { get; set; }
         public string Description { get; set; }
         public int Quantity { get; set; }
-        public double UnitPrice { get; set; }
-        public double Total { get; set; }
+        public decimal UnitPrice { get; set; }
+        public decimal Total { get; set; }
         public string? Note { get; set; }
     }
 }
